requirements.txt
fastapi
uvicorn
python-dateutil
pytz
pydantic
python-multipart
pyephem

datetime
flask==2.3.3
flask-cors==4.0.0

pyswisseph
